Fixed file herarchy, put player script in correct location
