--- Assets/Assets/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 86e380a21cf11674cb6bf61112720af8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
