--- Assets/Assets/Player/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85773366c7d8b0c4392a9f5aa5eeeaf8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
