+++ Assets/Assets/Scripts/Player/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4f6921014c7288242b60513918452c94
+guid: 85773366c7d8b0c4392a9f5aa5eeeaf8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
